Refactor: Create Layers
일단 레이어 생성부터 하고 보자
