+++ src/application/AppConfigurations.js
@@ -1,0 +1,5 @@
+export default class $AppConfigurations {
+  constructor() {
+    this.apiBaseURL = ''
+  }
+}
