+++ src/application/Store.js
@@ -1,0 +1,3 @@
+export default class $Store {
+
+}
