+++ src/util/APIs/Issue.js
@@ -1,4 +1,4 @@
-import {getOptions, customFetch} from "../MyFetch";
+import {getOptions, customFetch} from "../../data/repositories/utils/MyFetch.js";
 import {issue} from "../../models/api/RestAPIs";
 
 export const getIssue = customFetch(getOptions)(issue.get)
