+++ src/util/APIs/Label.js
@@ -1,4 +1,4 @@
-import {getOptions, customFetch} from "../MyFetch";
+import {getOptions, customFetch} from "../../data/repositories/utils/MyFetch.js";
 import {label} from "../../models/api/RestAPIs";
 
 export const getLabel = customFetch(getOptions)(label.get)
